fix confirmed bugs: parser norm order, wav chunk overflow, resampler truncation, session ctx clamp, ipc error handling, memory pressure race
- bug 36: re-trim whitespace after strip_trailing_punct in parse_command
- bug 37: use uint64_t for WAV unknown chunk skip to prevent overflow
- bug 38: ceiling division in to_mono prevents trailing sample loss
- bug 39: ceil() in resample_channel prevents last-sample truncation
- bug 40: clamp effective_ctx >= SYSTEM_PROMPT_TOKENS_RESERVED+1 in session
- bug 43: catch(...) in inference thread now sets inference_error_
- bug 44: memory_order_release/acquire on GCD memory pressure path (ARM safety)

+++ engine/src/audio/reader.cpp
@@ -143,7 +143,8 @@ WavFmt scan_wav_header(std::ifstream& f) {
         } else {
             // ── unknown chunk (LIST, INFO, fact, JUNK, …) ───────────────────
             // WAV spec: chunk payloads are padded to an even byte boundary.
-            uint32_t skip = chunk_size + (chunk_size & 1u);
+            // Use uint64_t to prevent overflow when chunk_size == 0xFFFFFFFF.
+            uint64_t skip = static_cast<uint64_t>(chunk_size) + (chunk_size & 1u);
             f.seekg(static_cast<std::streamoff>(skip), std::ios::cur);
         }
     }
